refactor(providerwire): split v4 request codec domains
